Extract fight movement into a separate module
- Move archer movement logic out of `scene.cpp` into `movement.cpp`
- Wire the new module into the test build
- Add a regression test for run-distance stability across tick rates
