+++ tests/CMakeLists.txt
@@ -16,8 +16,9 @@ enable_testing()
 add_executable(7dot_test
   archerCatalog.cpp
   fightCollision.cpp
+  fightMovement.cpp
   inputBuffer.cpp
-  level.cpp "../src/fight/collision.cpp" "../src/fight/level.cpp"
+  level.cpp "../src/fight/collision.cpp" "../src/fight/level.cpp" "../src/fight/movement.cpp"
   stageMapCatalog.cpp
   spriteCatalog.cpp
   "../src/renderer/ArcherCatalog.cpp"
